+++ scripts/generate-icons.ps1
@@ -59,5 +59,55 @@ function Write-TapirIcon([int]$size, [string]$fileName) {
   $bitmap.Dispose()
 }
 
+function Write-TapirWindowsIcon([string]$fileName) {
+  $source = [System.Drawing.Image]::FromFile((Join-Path $buildDir "icon.png"))
+  $sizes = @(16, 24, 32, 48, 64, 128, 256)
+  $images = @()
+
+  foreach ($size in $sizes) {
+    $bitmap = New-Object System.Drawing.Bitmap($size, $size, [System.Drawing.Imaging.PixelFormat]::Format32bppArgb)
+    $graphics = [System.Drawing.Graphics]::FromImage($bitmap)
+    $graphics.InterpolationMode = [System.Drawing.Drawing2D.InterpolationMode]::HighQualityBicubic
+    $graphics.SmoothingMode = [System.Drawing.Drawing2D.SmoothingMode]::AntiAlias
+    $graphics.PixelOffsetMode = [System.Drawing.Drawing2D.PixelOffsetMode]::HighQuality
+    $graphics.DrawImage($source, 0, 0, $size, $size)
+    $memory = New-Object System.IO.MemoryStream
+    $bitmap.Save($memory, [System.Drawing.Imaging.ImageFormat]::Png)
+    $images += ,@($size, $memory.ToArray())
+    $memory.Dispose()
+    $graphics.Dispose()
+    $bitmap.Dispose()
+  }
+  $source.Dispose()
+
+  $stream = [System.IO.File]::Create((Join-Path $buildDir $fileName))
+  $writer = New-Object System.IO.BinaryWriter($stream)
+  $writer.Write([uint16]0)
+  $writer.Write([uint16]1)
+  $writer.Write([uint16]$images.Count)
+  $offset = 6 + (16 * $images.Count)
+
+  foreach ($image in $images) {
+    $size = [int]$image[0]
+    $bytes = [byte[]]$image[1]
+    $writer.Write([byte]$(if ($size -eq 256) { 0 } else { $size }))
+    $writer.Write([byte]$(if ($size -eq 256) { 0 } else { $size }))
+    $writer.Write([byte]0)
+    $writer.Write([byte]0)
+    $writer.Write([uint16]1)
+    $writer.Write([uint16]32)
+    $writer.Write([uint32]$bytes.Length)
+    $writer.Write([uint32]$offset)
+    $offset += $bytes.Length
+  }
+
+  foreach ($image in $images) {
+    $writer.Write([byte[]]$image[1])
+  }
+  $writer.Dispose()
+  $stream.Dispose()
+}
+
 Write-TapirIcon 512 "icon.png"
 Write-TapirIcon 32 "tray-icon.png"
+Write-TapirWindowsIcon "icon.ico"